Time: 0 ms (100%), Space: 6 MB (71.97%) - LeetHub

+++ 1461-count-all-valid-pickup-and-delivery-options/1461-count-all-valid-pickup-and-delivery-options.cpp
@@ -1,15 +1,16 @@
 class Solution {
-    const int MOD = 1e9+7;
+
+    const int MOD = 1e9+7;  // used for the find the min number  
 public:
     int countOrders(int n) {
         // Using of Dp 
 
-        long long count = 1;
+        long long count = 1; // Create a variable 
         for(int i=2; i<=n; i++)
         {
             count = ( count * (2*i-1) * i)% MOD;
         }
-        return (int)count;
+        return count;
 
     }
 };
